chore: rename stringtable loaders to match style between games
