+++ src/ObjLoading/Game/IW3/StringTable/LoaderStringTableIW3.cpp
@@ -1,9 +1,9 @@
-#include "AssetLoaderStringTableT4.h"
+#include "LoaderStringTableIW3.h"
 
-#include "Game/T4/T4.h"
+#include "Game/IW3/IW3.h"
 #include "StringTable/StringTableLoader.h"
 
-using namespace T4;
+using namespace IW3;
 
 namespace
 {
@@ -38,7 +38,7 @@ namespace
 
 namespace string_table
 {
-    std::unique_ptr<AssetCreator<AssetStringTable>> CreateLoaderT4(MemoryManager& memory, ISearchPath& searchPath)
+    std::unique_ptr<AssetCreator<AssetStringTable>> CreateLoaderIW3(MemoryManager& memory, ISearchPath& searchPath)
     {
         return std::make_unique<StringTableLoader>(memory, searchPath);
     }
