+++ src/ObjLoading/Game/IW3/StringTable/LoaderStringTableIW3.h
