+++ src/ObjLoading/Game/T4/ObjLoaderT4.cpp
@@ -14,7 +14,7 @@
 #include "PhysPreset/GdtLoaderPhysPresetT4.h"
 #include "PhysPreset/RawLoaderPhysPresetT4.h"
 #include "RawFile/AssetLoaderRawFileT4.h"
-#include "StringTable/AssetLoaderStringTableT4.h"
+#include "StringTable/LoaderStringTableT4.h"
 #include "Weapon/FlameTableLoaderT4.h"
 #include "Weapon/WeaponGdtLoaderT4.h"
 #include "Weapon/WeaponRawLoaderT4.h"
